requirements.txt
Chroma==0.2.0
chromadb==0.3.21
python-dotenv==1.0.0
llama-index==0.5.27
